refactor: split revx-daemon lib.rs into ibc/mcp_render/mcp_tools
Mechanical move per #33, no behavior change:
- ibc.rs: CapabilityService::run_ibc_status/run_ibc_advance (356 lines)
- mcp_render.rs: mcp_response_summary + render/format helpers (1311 lines)
- mcp_tools.rs: tool_name_to_request + parse_empty_or_default + mcp_tools_manifest + tool_manifest (774 lines)
- lib.rs 10574 -> 8145 lines; moved items pub(crate), children glob-import parent
