+++ crates/revx-daemon/src/lib.rs
@@ -43,8 +43,15 @@ use tokio::io::{AsyncBufReadExt, AsyncWriteExt, BufReader};
 use tokio::net::TcpStream;
 use tracing::warn;
 
+mod ibc;
 mod mcp_http;
+mod mcp_render;
+mod mcp_tools;
+
 pub use mcp_http::serve_mcp_http;
+
+use mcp_render::*;
+use mcp_tools::*;
 
 const IBC_CONTINUUM_LEDGER_CACHE: &str = "ibc_continuum_ledger.json";
 
@@ -502,359 +509,6 @@ impl CapabilityService {
         drop(guard);
         self.persist_ibc_ledger(ws);
         bound_ids
-    }
-
-    fn run_ibc_status(&self, request: IbcStatusRequest) -> Result<IbcStatusResponse> {
-        let ws = self.workspace()?;
-        self.ensure_ibc_ledger_loaded(&ws);
-        let namespace = self.continuum_namespace(&ws, request.namespace.as_deref());
-        let (summary, focus, pc, status, epoch, witnesses, hyp_ids, lattice, next_actions) = {
-            let guard = self
-                .ibc_ledger
-                .lock()
-                .map_err(|_| anyhow::anyhow!("ibc ledger lock poisoned"))?;
-            let summary = revx_analysis::continuum_ledger_summary(&guard);
-            if let Some(state) = guard.sessions.get(&namespace) {
-                let plan = revx_analysis::lattice_ibc_plan(&state.lattice, state.focus, 6);
-                (
-                    summary,
-                    format!("{}@0x{:x}", state.focus_name, state.focus),
-                    state.lattice.ibc_pc,
-                    if state.lattice.ibc_status.is_empty() {
-                        "ready".to_string()
-                    } else {
-                        state.lattice.ibc_status.clone()
-                    },
-                    state.epoch,
-                    state.witnesses.clone(),
-                    state.orbit_hypotheses.values().cloned().collect::<Vec<_>>(),
-                    Some(state.lattice.clone()),
-                    plan,
-                )
-            } else {
-                (
-                    summary,
-                    String::new(),
-                    0,
-                    "empty".to_string(),
-                    0,
-                    Vec::new(),
-                    Vec::new(),
-                    None,
-                    Vec::new(),
-                )
-            }
-        };
-        let mut key_findings = vec![
-            summary.clone(),
-            format!("continuum_ns: {namespace}"),
-            format!("ibc_status: pc={pc} status={status} epoch={epoch}"),
-        ];
-        if !focus.is_empty() {
-            key_findings.push(format!("focus: {focus}"));
-        }
-        for w in witnesses.iter().rev().take(6) {
-            key_findings.push(format!("ibc_witness: {w}"));
-        }
-        if !hyp_ids.is_empty() {
-            key_findings.push(format!(
-                "orbit_hypotheses: {}",
-                hyp_ids
-                    .iter()
-                    .take(8)
-                    .cloned()
-                    .collect::<Vec<_>>()
-                    .join(",")
-            ));
-        }
-        if let Ok(guard) = self.ibc_ledger.lock()
-            && let Some(state) = guard.sessions.get(&namespace)
-        {
-            for line in revx_analysis::format_proof_chain_lines(&state.cognitive_field.proof_chain)
-                .into_iter()
-                .take(5)
-            {
-                key_findings.push(line);
-            }
-            let sealed = state
-                .cognitive_field
-                .proof_chain
-                .iter()
-                .filter(|l| l.verdict == "true" || l.verdict == "false")
-                .count();
-            if sealed > 0 {
-                key_findings.push(format!("pcos_sealed_links: {sealed}"));
-            }
-        }
-        let mut next = next_actions;
-        next.insert(
-            0,
-            AgentNextAction {
-                tool: "ibc_advance".to_string(),
-                reason: "advance IBC continuum one step (or warp via tool+query)".to_string(),
-                priority: 99,
-                query: None,
-                label: Some("ibc-advance".to_string()),
-                args: serde_json::json!({
-                    "namespace": namespace,
-                    "force_next": true
-                }),
-            },
-        );
-        if let Some(id) = hyp_ids.first() {
-            next.push(AgentNextAction {
-                tool: "hypothesis_update".to_string(),
-                reason: "edit confutation/status of bound orbit hypothesis".to_string(),
-                priority: 80,
-                query: Some(id.clone()),
-                label: Some("orbit-hyp".to_string()),
-                args: serde_json::json!({ "id": id }),
-            });
-        }
-        let agent_brief = AgentInteractionBrief {
-            headline: format!("IBC status ns={namespace} pc={pc} status={status}"),
-            key_findings,
-            open_questions: vec![
-                "Execute ibc_advance force_next or run the recommended tool+query".to_string(),
-            ],
-            next_actions: next.clone(),
-            stop_conditions: vec![
-                "Stop when PCOS proof_chain seals top orbits (true/false) and open residuals=0"
-                    .to_string(),
-                "Prefer residual probes over re-decompile of sealed orbits".to_string(),
-            ],
-            semantic_lattice: lattice.clone(),
-        };
-        Ok(IbcStatusResponse {
-            summary,
-            active_namespace: namespace,
-            focus,
-            pc,
-            status,
-            epoch,
-            witnesses,
-            hypothesis_ids: hyp_ids,
-            next_actions: next,
-            agent_brief,
-            semantic_lattice: lattice,
-        })
-    }
-
-    fn run_ibc_advance(&self, request: IbcAdvanceRequest) -> Result<IbcAdvanceResponse> {
-        let ws = self.workspace()?;
-        self.ensure_ibc_ledger_loaded(&ws);
-        let namespace = self.continuum_namespace(&ws, request.namespace.as_deref());
-        let tool = request
-            .tool
-            .clone()
-            .unwrap_or_else(|| "function_profile".to_string());
-        let (advanced, note, pc, status, epoch, hyp_ids, lattice, next_actions) = {
-            let mut guard = self
-                .ibc_ledger
-                .lock()
-                .map_err(|_| anyhow::anyhow!("ibc ledger lock poisoned"))?;
-            if !guard.sessions.contains_key(&namespace) {
-                guard.sessions.insert(
-                    namespace.clone(),
-                    revx_analysis::IbcContinuumState {
-                        namespace: namespace.clone(),
-                        focus: 0,
-                        focus_name: "unknown".to_string(),
-                        lattice: revx_core::AgentSemanticLattice::default(),
-                        witnesses: Vec::new(),
-                        orbit_hypotheses: std::collections::BTreeMap::new(),
-                        cognitive_field: revx_analysis::CognitiveField::default(),
-                        epoch: 0,
-                        updated_unix_ms: 0,
-                    },
-                );
-            }
-            let query_default_focus = guard.sessions.get(&namespace).map(|s| s.focus).unwrap_or(0);
-            let query = request
-                .query
-                .clone()
-                .unwrap_or_else(|| format!("0x{query_default_focus:x}"));
-            let mut global_w = None;
-            let result = {
-                let state = guard.sessions.get_mut(&namespace).unwrap();
-                let step = if request.force_next {
-                    revx_analysis::force_advance_ibc(&mut state.lattice)
-                } else {
-                    revx_analysis::observe_ibc_execution(&mut state.lattice, &tool, &query)
-                };
-                let corpus = revx_analysis::synthesize_observation_corpus(&state.lattice, None);
-                let collapse = revx_analysis::collapse_cognitive_field(
-                    &mut state.cognitive_field,
-                    &mut state.lattice,
-                    &tool,
-                    &query,
-                    Some(corpus.as_str()),
-                );
-                for event in &collapse {
-                    state
-                        .witnesses
-                        .push(format!("[{namespace}] COLLAPSE {event}"));
-                }
-                let mut field = revx_analysis::project_cognitive_field(&state.lattice);
-                field.field_epoch = state.epoch;
-                field.collapse_events = state.cognitive_field.collapse_events.clone();
-                field.collapse_events.extend(collapse.iter().cloned());
-                if field.collapse_events.len() > 24 {
-                    let n = field.collapse_events.len() - 24;
-                    field.collapse_events.drain(0..n);
-                }
-                field.residuals =
-                    revx_analysis::project_diffraction_residuals(&field, &state.lattice);
-                revx_analysis::apply_cognitive_field_to_lattice(&mut state.lattice, &field);
-                revx_analysis::inject_diffraction_residuals_into_lattice(
-                    &mut state.lattice,
-                    &field,
-                );
-                state.cognitive_field = field;
-                state.cognitive_field.proof_chain = revx_analysis::compose_proof_chain(state);
-                revx_analysis::inject_proof_chain_into_lattice(
-                    &mut state.lattice,
-                    &state.cognitive_field,
-                );
-                let advanced = step.is_some() || !collapse.is_empty();
-                if let Some(step) = step.as_ref() {
-                    state.epoch = state.epoch.saturating_add(1);
-                    let w = format!(
-                        "[{namespace}] {} {} => IBC[{}] {} | {}",
-                        tool, query, step.pc, step.op, step.detail
-                    );
-                    state.witnesses.push(w.clone());
-                    if state.witnesses.len() > 48 {
-                        let n = state.witnesses.len() - 48;
-                        state.witnesses.drain(0..n);
-                    }
-                    global_w = Some(w);
-                }
-                let note = if advanced {
-                    format!(
-                        "IBC continuum ADVANCED ns={namespace} epoch={} pc={}",
-                        state.epoch, state.lattice.ibc_pc
-                    )
-                } else {
-                    format!(
-                        "IBC continuum idle ns={namespace} pc={} status={}",
-                        state.lattice.ibc_pc,
-                        if state.lattice.ibc_status.is_empty() {
-                            "ready"
-                        } else {
-                            state.lattice.ibc_status.as_str()
-                        }
-                    )
-                };
-                let plan = revx_analysis::lattice_ibc_plan(&state.lattice, state.focus, 6);
-                (
-                    advanced,
-                    note,
-                    state.lattice.ibc_pc,
-                    if state.lattice.ibc_status.is_empty() {
-                        "ready".to_string()
-                    } else {
-                        state.lattice.ibc_status.clone()
-                    },
-                    state.epoch,
-                    state.orbit_hypotheses.values().cloned().collect::<Vec<_>>(),
-                    Some(state.lattice.clone()),
-                    plan,
-                )
-            };
-            if let Some(w) = global_w {
-                guard.global_witnesses.push(w);
-                if guard.global_witnesses.len() > 64 {
-                    let n = guard.global_witnesses.len() - 64;
-                    guard.global_witnesses.drain(0..n);
-                }
-            }
-            guard.active_namespace = namespace.clone();
-            result
-        };
-        self.persist_ibc_ledger(&ws);
-        let hyp_ids = {
-            let mut ids = hyp_ids;
-            let forged = self.forge_and_bind_orbit_hypotheses(&ws, &namespace);
-            for id in forged {
-                if !ids.contains(&id) {
-                    ids.push(id);
-                }
-            }
-            let sealed = self.seal_orbit_hypotheses_from_collapses(&ws, &namespace);
-            for id in sealed {
-                if !ids.contains(&id) {
-                    ids.push(id);
-                }
-            }
-            ids
-        };
-        let mut next = next_actions;
-        next.insert(
-            0,
-            AgentNextAction {
-                tool: "ibc_status".to_string(),
-                reason: "inspect durable continuum after advance".to_string(),
-                priority: 95,
-                query: None,
-                label: Some("ibc-status".to_string()),
-                args: serde_json::json!({ "namespace": namespace }),
-            },
-        );
-        let field_lines = lattice
-            .as_ref()
-            .map(|l| {
-                let f = revx_analysis::project_cognitive_field(l);
-                format!(
-                    "odc_field: mode={} entropy={:.2} residuals={} collapses={}",
-                    f.mode,
-                    f.entropy,
-                    f.residuals.len(),
-                    f.collapse_events.len()
-                )
-            })
-            .unwrap_or_else(|| "odc_field: -".to_string());
-        let agent_brief = AgentInteractionBrief {
-            headline: format!("IBC advance ns={namespace} advanced={advanced} pc={pc}"),
-            key_findings: vec![
-                note.clone(),
-                format!("continuum_ns: {namespace}"),
-                format!("ibc_status: pc={pc} status={status} epoch={epoch}"),
-                field_lines,
-                format!(
-                    "orbit_hypotheses: {}",
-                    hyp_ids
-                        .iter()
-                        .take(8)
-                        .cloned()
-                        .collect::<Vec<_>>()
-                        .join(",")
-                ),
-                format!("pcos_hypotheses_touched: {}", hyp_ids.len()),
-            ],
-            open_questions: vec![if advanced {
-                "Continue with next_actions[0] to follow IBC plan".to_string()
-            } else {
-                "No step matched; supply tool+query for warp or force_next=true".to_string()
-            }],
-            next_actions: next.clone(),
-            stop_conditions: vec![
-                "Prefer ibc_status after each advance before branching".to_string(),
-            ],
-            semantic_lattice: lattice.clone(),
-        };
-        Ok(IbcAdvanceResponse {
-            advanced,
-            note,
-            namespace,
-            pc,
-            status,
-            epoch,
-            hypothesis_ids: hyp_ids,
-            next_actions: next,
-            agent_brief,
-            semantic_lattice: lattice,
-        })
     }
 
     pub fn dispatch(&self, request: CapabilityRequest) -> Result<CapabilityResponse> {
@@ -2168,2089 +1822,6 @@ fn jsonrpc_error(id: serde_json::Value, code: i64, message: &str) -> serde_json:
             "code": code,
             "message": message,
         }
-    })
-}
-
-fn tool_name_to_request(name: &str, arguments: serde_json::Value) -> Result<CapabilityRequest> {
-    match name {
-        "project_open" => Ok(CapabilityRequest::ProjectOpen(serde_json::from_value(
-            arguments,
-        )?)),
-        "project_status" => Ok(CapabilityRequest::ProjectStatus(parse_empty_or_default(
-            arguments,
-        )?)),
-        "object_identify" => Ok(CapabilityRequest::ObjectIdentify(serde_json::from_value(
-            arguments,
-        )?)),
-        "object_search" => Ok(CapabilityRequest::ObjectSearch(serde_json::from_value(
-            arguments,
-        )?)),
-        "object_profile" => Ok(CapabilityRequest::ObjectProfile(serde_json::from_value(
-            arguments,
-        )?)),
-        "object_materialize" => Ok(CapabilityRequest::ObjectMaterialize(
-            serde_json::from_value(arguments)?,
-        )),
-        "object_extract_range" => Ok(CapabilityRequest::ObjectExtractRange(
-            serde_json::from_value(arguments)?,
-        )),
-        "object_scan_signatures" => Ok(CapabilityRequest::ObjectSignatureScan(
-            serde_json::from_value(arguments)?,
-        )),
-        "object_carve_signatures" => Ok(CapabilityRequest::ObjectCarveSignatures(
-            serde_json::from_value(arguments)?,
-        )),
-        "object_carve_identify" => Ok(CapabilityRequest::ObjectCarveIdentify(
-            serde_json::from_value(arguments)?,
-        )),
-        "object_analyze" => Ok(CapabilityRequest::ObjectAnalyze(serde_json::from_value(
-            arguments,
-        )?)),
-        "object_plugin_list" => Ok(CapabilityRequest::ObjectPluginList(parse_empty_or_default(
-            arguments,
-        )?)),
-        "object_plugin_run" => Ok(CapabilityRequest::ObjectPluginRun(serde_json::from_value(
-            arguments,
-        )?)),
-        "object_register_binary" => Ok(CapabilityRequest::ObjectRegisterBinary(
-            serde_json::from_value(arguments)?,
-        )),
-        "object_analyze_binary" => Ok(CapabilityRequest::ObjectAnalyzeBinary(
-            serde_json::from_value(arguments)?,
-        )),
-        "object_pipeline" => Ok(CapabilityRequest::ObjectPipeline(serde_json::from_value(
-            arguments,
-        )?)),
-        "binary_list" => Ok(CapabilityRequest::BinaryList(parse_empty_or_default(
-            arguments,
-        )?)),
-        "analysis_run" => Ok(CapabilityRequest::AnalysisRun(serde_json::from_value(
-            arguments,
-        )?)),
-        "analysis_status" => Ok(CapabilityRequest::AnalysisStatus(serde_json::from_value(
-            arguments,
-        )?)),
-        "binary_survey" => Ok(CapabilityRequest::BinarySurvey(serde_json::from_value(
-            arguments,
-        )?)),
-        "function_search" => Ok(CapabilityRequest::FunctionSearch(serde_json::from_value(
-            arguments,
-        )?)),
-        "function_profile" => Ok(CapabilityRequest::FunctionProfile(serde_json::from_value(
-            arguments,
-        )?)),
-        "decompile_function" => Ok(CapabilityRequest::DecompileFunction(
-            serde_json::from_value(arguments)?,
-        )),
-        "decompile_cache_status" => Ok(CapabilityRequest::DecompileCacheStatus(
-            serde_json::from_value(arguments)?,
-        )),
-        "disassemble_function" => Ok(CapabilityRequest::DisassembleFunction(
-            serde_json::from_value(arguments)?,
-        )),
-        "xrefs_query" => Ok(CapabilityRequest::XrefsQuery(serde_json::from_value(
-            arguments,
-        )?)),
-        "callgraph_slice" => Ok(CapabilityRequest::CallgraphSlice(serde_json::from_value(
-            arguments,
-        )?)),
-        "string_search" => Ok(CapabilityRequest::StringSearch(serde_json::from_value(
-            arguments,
-        )?)),
-        "search_bytes" => Ok(CapabilityRequest::SearchBytes(serde_json::from_value(
-            arguments,
-        )?)),
-        "object_search_content" => Ok(CapabilityRequest::ObjectContentSearch(
-            serde_json::from_value(arguments)?,
-        )),
-        "artifact_read" => Ok(CapabilityRequest::ArtifactRead(serde_json::from_value(
-            arguments,
-        )?)),
-        "artifact_list" => Ok(CapabilityRequest::ArtifactList(serde_json::from_value(
-            arguments,
-        )?)),
-        "evidence_pack" => Ok(CapabilityRequest::EvidencePack(serde_json::from_value(
-            arguments,
-        )?)),
-        "evidence_graph" => Ok(CapabilityRequest::EvidenceGraph(serde_json::from_value(
-            arguments,
-        )?)),
-        "symbolic_solve" => Ok(CapabilityRequest::SymbolicSolve(serde_json::from_value(
-            arguments,
-        )?)),
-        "analysis_brief" => Ok(CapabilityRequest::AnalysisBrief(serde_json::from_value(
-            arguments,
-        )?)),
-        "investigation_run" => Ok(CapabilityRequest::InvestigationRun(serde_json::from_value(
-            arguments,
-        )?)),
-        "ibc_status" => Ok(CapabilityRequest::IbcStatus(parse_empty_or_default(
-            arguments,
-        )?)),
-        "ibc_advance" => Ok(CapabilityRequest::IbcAdvance(parse_empty_or_default(
-            arguments,
-        )?)),
-        "hypothesis_create" => Ok(CapabilityRequest::HypothesisCreate(serde_json::from_value(
-            arguments,
-        )?)),
-        "hypothesis_update" => Ok(CapabilityRequest::HypothesisUpdate(serde_json::from_value(
-            arguments,
-        )?)),
-        "report_generate" => Ok(CapabilityRequest::ReportGenerate(serde_json::from_value(
-            arguments,
-        )?)),
-        "trace_import" => Ok(CapabilityRequest::TraceImport(serde_json::from_value(
-            arguments,
-        )?)),
-        "trace_query" => Ok(CapabilityRequest::TraceQuery(serde_json::from_value(
-            arguments,
-        )?)),
-        _ => anyhow::bail!("unknown tool: {name}"),
-    }
-}
-
-fn parse_empty_or_default<T>(arguments: serde_json::Value) -> Result<T>
-where
-    T: serde::de::DeserializeOwned + Default,
-{
-    match arguments {
-        serde_json::Value::Null => Ok(T::default()),
-        serde_json::Value::Object(map) if map.is_empty() => Ok(T::default()),
-        other => Ok(serde_json::from_value(other)?),
-    }
-}
-
-fn mcp_response_summary(response: &CapabilityResponse) -> String {
-    const MAX_CHARS: usize = 24_000;
-    let body = match response {
-        CapabilityResponse::ProjectOpen(payload) => format!(
-            "# project_open\nworkspace: {}\nproject: {}\nschema: {}",
-            payload.workspace_root, payload.project.name, payload.project.schema_version
-        ),
-        CapabilityResponse::ProjectStatus(payload) => {
-            let mut lines = vec![
-                "# project_status".to_string(),
-                format!("workspace: {}", payload.workspace_root),
-                format!("project: {}", payload.project.name),
-                format!("binaries: {}", payload.binary_count),
-            ];
-            if !payload.binaries.is_empty() {
-                lines.push("\n## Binaries".to_string());
-                for binary in payload.binaries.iter().take(40) {
-                    lines.push(format!(
-                        "- {}  path={}  {:?}/{:?}  funcs={} imports={} strings={} typed={} pseudocode={}",
-                        binary.id,
-                        binary.path,
-                        binary.format,
-                        binary.architecture,
-                        binary.function_count,
-                        binary.import_count,
-                        binary.string_count,
-                        binary.typed_function_count,
-                        binary.structured_pseudocode_count
-                    ));
-                }
-                if payload.binaries.len() > 40 {
-                    lines.push(format!("- ... {} more", payload.binaries.len() - 40));
-                }
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ObjectIdentify(payload) => {
-            let mut lines = vec![
-                "# object_identify".to_string(),
-                format!("root: {}", payload.root_id),
-                format!(
-                    "objects: {}  edges: {}  evidence: {}",
-                    payload.object_count, payload.edge_count, payload.evidence_count
-                ),
-            ];
-            if let Some(graph) = &payload.graph {
-                lines.push("\n## Objects".to_string());
-                for object in graph.objects.iter().take(30) {
-                    lines.push(format!(
-                        "- {}  name={}  kind={:?}  format={}  size={}  depth={}",
-                        object.id,
-                        object.display_name,
-                        object.kind,
-                        object.format.as_deref().unwrap_or("-"),
-                        object.size,
-                        object.depth
-                    ));
-                }
-                if graph.objects.len() > 30 {
-                    lines.push(format!("- ... {} more objects", graph.objects.len() - 30));
-                }
-            }
-            if !payload.evidence_ids.is_empty() {
-                lines.push(format!(
-                    "\n## Evidence IDs\n{}",
-                    format_id_list(&payload.evidence_ids, 20)
-                ));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ObjectSearch(payload) => {
-            let mut lines = vec![
-                "# object_search".to_string(),
-                format!("matches: {}", payload.objects.len()),
-            ];
-            for object in payload.objects.iter().take(40) {
-                lines.push(format!(
-                    "- {}  name={}  kind={:?}  format={}  size={}",
-                    object.id,
-                    object.display_name,
-                    object.kind,
-                    object.format.as_deref().unwrap_or("-"),
-                    object.size
-                ));
-            }
-            if payload.objects.len() > 40 {
-                lines.push(format!("- ... {} more", payload.objects.len() - 40));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ObjectProfile(payload) => {
-            let mut lines = vec![
-                "# object_profile".to_string(),
-                format!(
-                    "id: {}  name: {}  kind: {:?}  format: {}  size: {}",
-                    payload.object.id,
-                    payload.object.display_name,
-                    payload.object.kind,
-                    payload.object.format.as_deref().unwrap_or("-"),
-                    payload.object.size
-                ),
-            ];
-            if !payload.incoming_edges.is_empty() {
-                lines.push("\n## Incoming".to_string());
-                for edge in payload.incoming_edges.iter().take(20) {
-                    lines.push(format!("- {:?} {} -> {}", edge.kind, edge.from, edge.to));
-                }
-            }
-            if !payload.outgoing_edges.is_empty() {
-                lines.push("\n## Outgoing".to_string());
-                for edge in payload.outgoing_edges.iter().take(20) {
-                    lines.push(format!("- {:?} {} -> {}", edge.kind, edge.from, edge.to));
-                }
-            }
-            if !payload.object.analyses.is_empty() {
-                lines.push("\n## Analyses".to_string());
-                for analysis in payload.object.analyses.iter().take(12) {
-                    lines.push(format!(
-                        "- {} [{:?}]: {}",
-                        analysis.analyzer,
-                        analysis.status,
-                        truncate_chars(&analysis.summary, 240)
-                    ));
-                }
-            }
-            if !payload.evidence_ids.is_empty() {
-                lines.push(format!(
-                    "\n## Evidence IDs\n{}",
-                    format_id_list(&payload.evidence_ids, 20)
-                ));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ObjectMaterialize(payload) => format!(
-            "# object_materialize\nid: {}\nname: {}\nartifact: {}\nsize: {}\ncontent_type: {}\nevidence: {}\npreview_hex: {}\npreview_text:\n{}",
-            payload.object.id,
-            payload.object.display_name,
-            payload.artifact.relative_path,
-            payload.artifact.size,
-            payload.artifact.content_type,
-            payload.evidence_id,
-            truncate_chars(payload.preview_hex.as_deref().unwrap_or(""), 240),
-            truncate_chars(payload.preview_text.as_deref().unwrap_or(""), 1_500)
-        ),
-        CapabilityResponse::ObjectExtractRange(payload) => format!(
-            "# object_extract_range\nid: {}\noffset: 0x{:x}\nrequested: {}\nextracted: {}\nartifact: {}\nevidence: {}\npreview_hex: {}\npreview_text:\n{}",
-            payload.object.id,
-            payload.offset,
-            payload.requested_length,
-            payload.extracted_size,
-            payload.artifact.relative_path,
-            payload.evidence_id,
-            truncate_chars(payload.preview_hex.as_deref().unwrap_or(""), 240),
-            truncate_chars(payload.preview_text.as_deref().unwrap_or(""), 1_500)
-        ),
-        CapabilityResponse::ObjectSignatureScan(payload) => {
-            let mut lines = vec![
-                "# object_scan_signatures".to_string(),
-                format!(
-                    "object: {}  scanned: {}  signatures: {}  truncated: {}",
-                    payload.object.id,
-                    payload.scanned_size,
-                    payload.returned_count,
-                    payload.truncated
-                ),
-                format!("artifact: {}", payload.artifact.relative_path),
-                format!("evidence: {}", payload.evidence_id),
-            ];
-            if !payload.signatures.is_empty() {
-                lines.push("\n## Signatures".to_string());
-                for hit in payload.signatures.iter().take(30) {
-                    lines.push(format!(
-                        "- 0x{:x}  {}  format={}  conf={:.2}  len={:?}  {}",
-                        hit.offset,
-                        hit.signature,
-                        hit.format,
-                        hit.confidence,
-                        hit.suggested_length,
-                        truncate_chars(&hit.description, 120)
-                    ));
-                }
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ObjectCarveSignatures(payload) => {
-            let mut lines = vec![
-                "# object_carve_signatures".to_string(),
-                format!(
-                    "object: {}  scanned: {}  carved: {}  skipped: {}  truncated: {}",
-                    payload.object.id,
-                    payload.scanned_count,
-                    payload.carved_count,
-                    payload.skipped_count,
-                    payload.truncated
-                ),
-                format!("artifact: {}", payload.artifact.relative_path),
-                format!("evidence: {}", payload.carve_evidence_id),
-            ];
-            if !payload.carves.is_empty() {
-                lines.push("\n## Carves".to_string());
-                for carve in payload.carves.iter().take(20) {
-                    lines.push(format!(
-                        "- 0x{:x}+{}  {}  format={}  conf={:.2}  artifact={}",
-                        carve.offset,
-                        carve.length,
-                        carve.signature,
-                        carve.format,
-                        carve.confidence,
-                        carve.artifact.relative_path
-                    ));
-                }
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ObjectCarveIdentify(payload) => {
-            let mut lines = vec![
-                "# object_carve_identify".to_string(),
-                format!(
-                    "object: {}  carved: {}  identified: {}  failed: {}",
-                    payload.object.id,
-                    payload.carved_count,
-                    payload.identified_count,
-                    payload.failed_count
-                ),
-                format!("artifact: {}", payload.artifact.relative_path),
-            ];
-            for item in payload.carves.iter().take(20) {
-                lines.push(format!(
-                    "- carve 0x{:x}+{} -> root={} objects={} edges={} error={}",
-                    item.carve.offset,
-                    item.carve.length,
-                    item.root_id.as_deref().unwrap_or("-"),
-                    item.object_count,
-                    item.edge_count,
-                    item.error.as_deref().unwrap_or("-")
-                ));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ObjectAnalyze(payload) => {
-            let mut lines = vec![
-                "# object_analyze".to_string(),
-                format!(
-                    "object: {}  name: {}  kind: {:?}  format: {}",
-                    payload.object.id,
-                    payload.object.display_name,
-                    payload.object.kind,
-                    payload.object.format.as_deref().unwrap_or("-")
-                ),
-                format_agent_brief_section(&payload.agent_brief),
-            ];
-            if !payload.analyses.is_empty() {
-                lines.push("\n## Analyses".to_string());
-                for analysis in payload.analyses.iter().take(16) {
-                    lines.push(format!(
-                        "- {} [{:?}]: {}",
-                        analysis.analyzer,
-                        analysis.status,
-                        truncate_chars(&analysis.summary, 320)
-                    ));
-                }
-            }
-            if !payload.evidence_ids.is_empty() {
-                lines.push(format!(
-                    "\n## Evidence IDs\n{}",
-                    format_id_list(&payload.evidence_ids, 20)
-                ));
-            }
-            if let Some(artifact) = &payload.artifact {
-                lines.push(format!("\nartifact: {}", artifact.relative_path));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ObjectPluginList(payload) => {
-            let mut lines = vec![
-                "# object_plugin_list".to_string(),
-                format!("plugins: {}", payload.plugins.len()),
-            ];
-            for plugin in payload.plugins.iter().take(40) {
-                lines.push(format!(
-                    "- {}  {}  timeout_ms={:?}",
-                    plugin.id,
-                    plugin
-                        .description
-                        .as_deref()
-                        .unwrap_or(plugin.name.as_str()),
-                    plugin.timeout_ms
-                ));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ObjectPluginRun(payload) => {
-            let mut lines = vec![
-                "# object_plugin_run".to_string(),
-                format!(
-                    "plugin: {}  object: {}  status: {:?}  evidence: {}",
-                    payload.plugin.id, payload.object.id, payload.status, payload.evidence_id
-                ),
-                format!("summary: {}", truncate_chars(&payload.summary, 500)),
-            ];
-            if let Some(stdout) = &payload.stdout_preview {
-                lines.push(format!("\n## stdout\n{}", truncate_chars(stdout, 2_000)));
-            }
-            if let Some(stderr) = &payload.stderr_preview {
-                lines.push(format!("\n## stderr\n{}", truncate_chars(stderr, 1_000)));
-            }
-            if let Some(json) = &payload.output_json {
-                lines.push(format!(
-                    "\n## output_json\n{}",
-                    truncate_chars(&json.to_string(), 2_000)
-                ));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ObjectRegisterBinary(payload) => format!(
-            "# object_register_binary\nobject: {}\nbinary: {}\npath: {}\nformat: {:?}\narch: {:?}\nentry: {:?}\nfunctions: {}\nimports: {}\nexports: {}\nstrings: {}\nevidence: {}\nsurvey_artifact: {}",
-            payload.object.id,
-            payload.survey.binary.id,
-            payload.survey.binary.path,
-            payload.survey.binary.format,
-            payload.survey.binary.architecture,
-            payload.survey.binary.entry,
-            payload.survey.summary.function_count,
-            payload.survey.summary.import_count,
-            payload.survey.summary.export_count,
-            payload.survey.summary.string_count,
-            payload.evidence_id,
-            payload.survey_artifact.relative_path
-        ),
-        CapabilityResponse::ObjectAnalyzeBinary(payload) => format!(
-            "# object_analyze_binary\nobject: {}\nrun_id: {}\nstatus: {:?}\n{}\nevidence: {}\n{}",
-            payload.object.id,
-            payload.run_id,
-            payload.status,
-            format_analysis_summary(&payload.summary),
-            payload.evidence_count,
-            format_id_list(&payload.evidence_ids, 16)
-        ),
-        CapabilityResponse::ObjectPipeline(payload) => {
-            let mut lines = vec![
-                "# object_pipeline".to_string(),
-                format!("pipeline_id: {}", payload.pipeline_id),
-                format!(
-                    "root: {}  objects: {}  edges: {}  analyzed_objects: {}  carved: {}  embedded: {}  binaries: {}  failed: {}  evidence: {}",
-                    payload.root_id,
-                    payload.object_count,
-                    payload.edge_count,
-                    payload.analyzed_object_count,
-                    payload.carved_object_count,
-                    payload.identified_embedded_object_count,
-                    payload.analyzed_binary_count,
-                    payload.failed_step_count,
-                    payload.evidence_count
-                ),
-                format_agent_brief_section(&payload.agent_brief),
-            ];
-            if !payload.steps.is_empty() {
-                lines.push("\n## Steps".to_string());
-                for step in payload.steps.iter().take(24) {
-                    lines.push(format!(
-                        "- {:?} {:?} [{:?}]: {}",
-                        step.stage,
-                        step.object_path,
-                        step.status,
-                        truncate_chars(&step.summary, 200)
-                    ));
-                }
-                if payload.steps.len() > 24 {
-                    lines.push(format!("- ... {} more steps", payload.steps.len() - 24));
-                }
-            }
-            lines.push(format!(
-                "\nreport_artifact: {}\ngraph_artifact: {}",
-                payload.report_artifact.relative_path, payload.graph_artifact.relative_path
-            ));
-            lines.join("\n")
-        }
-        CapabilityResponse::BinaryList(payload) => {
-            let mut lines = vec![
-                "# binary_list".to_string(),
-                format!("binaries: {}", payload.binaries.len()),
-            ];
-            for binary in payload.binaries.iter().take(40) {
-                lines.push(format!(
-                    "- {}  path={}  {:?}/{:?}  funcs={} imports={} exports={} strings={}",
-                    binary.id,
-                    binary.path,
-                    binary.format,
-                    binary.architecture,
-                    binary.function_count,
-                    binary.import_count,
-                    binary.export_count,
-                    binary.string_count
-                ));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::AnalysisRun(payload) => format!(
-            "# analysis_run\nrun_id: {}\nstatus: {:?}\n{}\nevidence_count: {}\nevidence_ids:\n{}\nnext: function_search / binary_survey / string_search",
-            payload.run_id,
-            payload.status,
-            format_analysis_summary(&payload.summary),
-            payload.evidence_count,
-            format_id_list(&payload.evidence_ids, 20)
-        ),
-        CapabilityResponse::AnalysisStatus(payload) => format!(
-            "# analysis_status\nrun_id: {}\nbinary_id: {}\nprofile: {:?}\nstatus: {:?}\ncreated_at: {}\ncompleted_at: {:?}\n{}",
-            payload.run_id,
-            payload.binary_id,
-            payload.profile,
-            payload.status,
-            payload.created_at,
-            payload.completed_at,
-            format_analysis_summary(&payload.summary)
-        ),
-        CapabilityResponse::BinarySurvey(payload) => {
-            let mut lines = vec![
-                "# binary_survey".to_string(),
-                format_analysis_summary(&payload.preview),
-                format!("evidence_count: {}", payload.evidence_count),
-            ];
-            if let Some(survey) = &payload.survey {
-                lines.push(format!(
-                    "path: {}\nentry: {:?}\nimage_base: {:?}\nsize: {}\nhash: {}",
-                    survey.binary.path,
-                    survey.binary.entry,
-                    survey.binary.image_base,
-                    survey.binary.size,
-                    survey.binary.hash_blake3
-                ));
-            }
-            if !payload.evidence_ids.is_empty() {
-                lines.push(format!(
-                    "\n## Evidence IDs\n{}",
-                    format_id_list(&payload.evidence_ids, 20)
-                ));
-            }
-            lines.push(
-                "\nnext: function_search(query) | string_search(pattern) | function_profile(name|addr)"
-                    .to_string(),
-            );
-            lines.join("\n")
-        }
-        CapabilityResponse::FunctionSearch(payload) => {
-            let mut lines = vec![
-                "# function_search".to_string(),
-                format!("matches: {}", payload.functions.len()),
-            ];
-            if payload.functions.is_empty() {
-                lines.push("no matches".to_string());
-            } else {
-                for function in payload.functions.iter().take(50) {
-                    lines.push(format!(
-                        "- {}  0x{:x}  size={}  evidence={}",
-                        function.name,
-                        function.address,
-                        function.size,
-                        function.evidence_ids.len()
-                    ));
-                }
-                if payload.functions.len() > 50 {
-                    lines.push(format!("- ... {} more", payload.functions.len() - 50));
-                }
-                lines.push(
-                    "\nnext: function_profile(query) | decompile_function(query) | disassemble_function(query)"
-                        .to_string(),
-                );
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::FunctionProfile(payload) => render_function_profile(payload),
-        CapabilityResponse::DecompileFunction(payload) => {
-            let mut lines = vec![
-                "# decompile_function".to_string(),
-                format!("name: {}", payload.function_name),
-                format!("address: 0x{:x}", payload.address),
-                format!("strategy: {:?}", payload.strategy_used),
-                format!("cache_hit: {}", payload.cache_hit),
-            ];
-            if !payload.available_strategies.is_empty() {
-                lines.push(format!(
-                    "available_strategies: {}",
-                    payload.available_strategies.join(",")
-                ));
-            }
-            if let Some(unit) = &payload.pseudocode {
-                if let Some(lattice) = &unit.semantic_lattice {
-                    lines.push(String::new());
-                    lines.push(revx_analysis::format_semantic_lattice(lattice));
-                } else if let Some(lattice) = &payload.agent_brief.semantic_lattice {
-                    lines.push(String::new());
-                    lines.push(revx_analysis::format_semantic_lattice(lattice));
-                }
-            } else if let Some(lattice) = &payload.agent_brief.semantic_lattice {
-                lines.push(String::new());
-                lines.push(revx_analysis::format_semantic_lattice(lattice));
-            }
-            lines.push(format!(
-                "\n## Digest\n{}",
-                function_pseudocode_digest(payload.pseudocode.as_ref(), &[], &[])
-            ));
-            match &payload.pseudocode {
-                Some(unit) => {
-                    lines.push(format!("language: {}", unit.language));
-                    lines.push(format!(
-                        "regions: {}  evidence: {}",
-                        unit.regions.len(),
-                        unit.evidence_ids.len()
-                    ));
-                    lines.push("\n## Pseudocode".to_string());
-                    lines.push(format!("```{}\n{}\n```", unit.language, unit.text));
-                    if !unit.regions.is_empty() {
-                        lines.push("\n## Regions".to_string());
-                        for region in unit.regions.iter().take(24) {
-                            lines.push(format!(
-                                "- {} {:?}  {:x?}-{:x?}  stmts={}  {}",
-                                region.id,
-                                region.kind,
-                                region.start_address,
-                                region.end_address,
-                                region.statements.len(),
-                                region.header.as_deref().unwrap_or("")
-                            ));
-                        }
-                    }
-                }
-                None => lines.push(
-                    "pseudocode: unavailable\nnext: disassemble_function(query) | function_profile(query)"
-                        .to_string(),
-                ),
-            }
-            if !payload.evidence_ids.is_empty() {
-                lines.push(format!(
-                    "\n## Evidence IDs\n{}",
-                    format_id_list(&payload.evidence_ids, 20)
-                ));
-            }
-            if !payload.agent_brief.headline.is_empty()
-                || !payload.agent_brief.next_actions.is_empty()
-            {
-                lines.push(format_agent_brief_section(&payload.agent_brief));
-            }
-            if let Some(artifact) = &payload.artifact {
-                lines.push(format!("\nartifact: {}", artifact.relative_path));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::DecompileCacheStatus(payload) => {
-            let mut lines = vec![
-                "# decompile_cache_status".to_string(),
-                format!("name: {}", payload.function_name),
-                format!("address: 0x{:x}", payload.address),
-                format!(
-                    "function_pseudocode: {} regions={} text_len={}",
-                    payload.has_function_pseudocode,
-                    payload.function_region_count,
-                    payload.function_text_len
-                ),
-                format!("strategy_caches: {}", payload.strategies.len()),
-            ];
-            if payload.strategies.is_empty() {
-                lines.push("no strategy cache entries".to_string());
-            } else {
-                for entry in &payload.strategies {
-                    lines.push(format!(
-                        "- {} regions={} text_len={} lattice={}",
-                        entry.strategy, entry.region_count, entry.text_len, entry.has_lattice
-                    ));
-                }
-            }
-            lines.push(
-                "
-next: decompile_function(query, strategy) | function_profile(query)"
-                    .to_string(),
-            );
-            lines.join(
-                "
-",
-            )
-        }
-        CapabilityResponse::DisassembleFunction(payload) => render_disassembly(payload),
-        CapabilityResponse::XrefsQuery(payload) => {
-            let mut lines = vec![
-                "# xrefs_query".to_string(),
-                format!("references: {}", payload.references.len()),
-            ];
-            for reference in payload.references.iter().take(80) {
-                lines.push(format!(
-                    "- 0x{:x} -> 0x{:x}  {}",
-                    reference.from, reference.to, reference.kind
-                ));
-            }
-            if payload.references.len() > 80 {
-                lines.push(format!("- ... {} more", payload.references.len() - 80));
-            }
-            if !payload.agent_brief.headline.is_empty()
-                || !payload.agent_brief.next_actions.is_empty()
-            {
-                lines.push(format_agent_brief_section(&payload.agent_brief));
-            } else if !payload.references.is_empty() {
-                lines.push(
-                    "\nnext: function_profile(0xaddr) | decompile_function(0xaddr) | disassemble_function(0xaddr)"
-                        .to_string(),
-                );
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::CallgraphSlice(payload) => {
-            let mut lines = vec![
-                "# callgraph_slice".to_string(),
-                format!("edges: {}", payload.edges.len()),
-            ];
-            for edge in payload.edges.iter().take(60) {
-                lines.push(format!(
-                    "- {} (0x{:x}) -[{}]-> {} (0x{:x})",
-                    edge.caller_name,
-                    edge.caller_address,
-                    edge.kind,
-                    edge.callee_name.as_deref().unwrap_or("?"),
-                    edge.callee_address
-                ));
-            }
-            if payload.edges.len() > 60 {
-                lines.push(format!("- ... {} more", payload.edges.len() - 60));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::StringSearch(payload) => {
-            let mut lines = vec![
-                "# string_search".to_string(),
-                format!("matches: {}", payload.matches.len()),
-            ];
-            for item in payload.matches.iter().take(60) {
-                match item.address {
-                    Some(address) => lines.push(format!(
-                        "- 0x{:x}  {}",
-                        address,
-                        truncate_chars(&item.value, 200)
-                    )),
-                    None => lines.push(format!("- {}", truncate_chars(&item.value, 200))),
-                }
-            }
-            if payload.matches.len() > 60 {
-                lines.push(format!("- ... {} more", payload.matches.len() - 60));
-            }
-            if !payload.agent_brief.headline.is_empty()
-                || !payload.agent_brief.next_actions.is_empty()
-            {
-                lines.push(format_agent_brief_section(&payload.agent_brief));
-            } else if !payload.matches.is_empty() {
-                lines.push(
-                    "\nnext: xrefs_query(0xstring_addr) then function_profile on owning function"
-                        .to_string(),
-                );
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::SearchBytes(payload) => {
-            let mut lines = vec![
-                "# search_bytes".to_string(),
-                format!("matches: {}", payload.matches.len()),
-            ];
-            for item in payload.matches.iter().take(40) {
-                lines.push(format!(
-                    "- 0x{:x}  {}",
-                    item.offset,
-                    truncate_chars(&item.bytes, 120)
-                ));
-            }
-            if payload.matches.len() > 40 {
-                lines.push(format!("- ... {} more", payload.matches.len() - 40));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ObjectContentSearch(payload) => {
-            let mut lines = vec![
-                "# object_search_content".to_string(),
-                format!(
-                    "pattern: {}  mode: {:?}  objects: {}  searched: {}  matches: {}  truncated: {}",
-                    payload.pattern,
-                    payload.mode,
-                    payload.object_count,
-                    payload.searched_object_count,
-                    payload.returned_count,
-                    payload.truncated
-                ),
-            ];
-            for item in payload.matches.iter().take(40) {
-                lines.push(format!(
-                    "- {}  0x{:x}+{}  kind={:?}  text={}",
-                    item.display_name,
-                    item.offset,
-                    item.length,
-                    item.object_kind,
-                    truncate_chars(
-                        item.preview_text.as_deref().unwrap_or(&item.preview_hex),
-                        160
-                    )
-                ));
-            }
-            if payload.matches.len() > 40 {
-                lines.push(format!("- ... {} more", payload.matches.len() - 40));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ArtifactRead(payload) => {
-            let mut lines = vec![
-                "# artifact_read".to_string(),
-                format!(
-                    "path: {}  hash: {}  offset: {}  returned: {}/{}  truncated: {}",
-                    payload.artifact.relative_path,
-                    payload.artifact.hash_blake3,
-                    payload.offset,
-                    payload.returned_size,
-                    payload.total_size,
-                    payload.truncated
-                ),
-                format!("content_type: {}", payload.artifact.content_type),
-            ];
-            if let Some(text) = &payload.preview_text {
-                lines.push(format!("\n## Text\n{}", truncate_chars(text, 8_000)));
-            } else {
-                lines.push(format!(
-                    "\n## Hex\n{}",
-                    truncate_chars(&payload.preview_hex, 2_000)
-                ));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::ArtifactList(payload) => {
-            let mut lines = vec![
-                "# artifact_list".to_string(),
-                format!(
-                    "returned: {}/{}  truncated: {}",
-                    payload.returned_count, payload.total_count, payload.truncated
-                ),
-            ];
-            for hit in payload.artifacts.iter().take(40) {
-                lines.push(format!(
-                    "- {}  {}  size={}  roles=[{}]  refs={}",
-                    hit.artifact.relative_path,
-                    hit.artifact.content_type,
-                    hit.artifact.size,
-                    hit.roles.join(","),
-                    hit.references.len()
-                ));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::EvidencePack(payload) => {
-            let mut lines = vec![
-                "# evidence_pack".to_string(),
-                format!("preview: {}", payload.preview.len()),
-            ];
-            for evidence in payload.preview.iter().take(40) {
-                lines.push(format!(
-                    "- [{}] {}  subject={}  {}",
-                    evidence.kind,
-                    evidence.id,
-                    evidence.subject,
-                    truncate_chars(&evidence.summary, 220)
-                ));
-                if !evidence.details.is_null() {
-                    let details = truncate_chars(&evidence.details.to_string(), 280);
-                    if details != "null" && details != "{}" {
-                        lines.push(format!("  details: {details}"));
-                    }
-                }
-            }
-            if payload.preview.len() > 40 {
-                lines.push(format!("- ... {} more", payload.preview.len() - 40));
-            }
-            if let Some(artifact) = &payload.artifact {
-                lines.push(format!("\nartifact: {}", artifact.relative_path));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::EvidenceGraph(payload) => {
-            let mut lines = vec![
-                "# evidence_graph".to_string(),
-                format!(
-                    "subject: {}  nodes: {}  edges: {}  evidence: {}",
-                    payload.subject, payload.node_count, payload.edge_count, payload.evidence_count
-                ),
-                format!("artifact: {}", payload.artifact.relative_path),
-            ];
-            if !payload.nodes.is_empty() {
-                lines.push("\n## Nodes".to_string());
-                for node in payload.nodes.iter().take(40) {
-                    lines.push(format!(
-                        "- {}  kind={}  {}  {}",
-                        node.id,
-                        node.kind,
-                        node.label,
-                        truncate_chars(node.summary.as_deref().unwrap_or(""), 120)
-                    ));
-                }
-            }
-            if !payload.edges.is_empty() {
-                lines.push("\n## Edges".to_string());
-                for edge in payload.edges.iter().take(40) {
-                    lines.push(format!(
-                        "- {} -[{}/{}]-> {}",
-                        edge.from, edge.kind, edge.label, edge.to
-                    ));
-                }
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::SymbolicSolve(payload) => {
-            let mut lines = vec![
-                "# symbolic_solve".to_string(),
-                format!(
-                    "case: {}  subject: {}  status: {:?}  constraints: {}  checked: {}",
-                    payload.case_id,
-                    payload.subject,
-                    payload.status,
-                    payload.constraint_count,
-                    payload.checked_assignments
-                ),
-                format!("evidence: {}", payload.evidence_id),
-            ];
-            if !payload.solutions.is_empty() {
-                lines.push("\n## Solutions".to_string());
-                for (index, solution) in payload.solutions.iter().take(10).enumerate() {
-                    let pairs = solution
-                        .iter()
-                        .map(|(k, v)| format!("{k}={v}"))
-                        .collect::<Vec<_>>()
-                        .join(", ");
-                    lines.push(format!("- #{index}: {pairs}"));
-                }
-            }
-            if !payload.warnings.is_empty() {
-                lines.push("\n## Warnings".to_string());
-                for warning in payload.warnings.iter().take(10) {
-                    lines.push(format!("- {warning}"));
-                }
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::AnalysisBrief(payload) => render_analysis_brief(payload),
-        CapabilityResponse::InvestigationRun(payload) => {
-            let mut lines = vec![
-                "# investigation_run".to_string(),
-                format!("id: {}", payload.investigation_id),
-                format!("subject: {}", payload.subject),
-                format!(
-                    "evidence: {}  graph_nodes: {}  graph_edges: {}  traces: {}",
-                    payload.evidence_count,
-                    payload.graph.node_count,
-                    payload.graph.edge_count,
-                    payload.trace_count
-                ),
-                format_agent_brief_section(&payload.agent_brief),
-                format!("\n## Summary\n{}", truncate_chars(&payload.summary, 2_000)),
-            ];
-            if !payload.report.body.is_empty() {
-                lines.push(format!(
-                    "\n## Report\n{}",
-                    truncate_chars(&payload.report.body, 8_000)
-                ));
-            }
-            lines.push(format!(
-                "\nreport_artifact: {}\nartifact: {}",
-                payload.report_artifact.relative_path, payload.artifact.relative_path
-            ));
-            lines.join("\n")
-        }
-        CapabilityResponse::IbcStatus(payload) => {
-            let mut lines = vec![
-                "# ibc_status".to_string(),
-                format!("namespace: {}", payload.active_namespace),
-                format!("focus: {}", payload.focus),
-                format!(
-                    "pc: {}  status: {}  epoch: {}",
-                    payload.pc, payload.status, payload.epoch
-                ),
-                format!("summary: {}", payload.summary),
-                format_agent_brief_section(&payload.agent_brief),
-            ];
-            if !payload.hypothesis_ids.is_empty() {
-                lines.push(format!(
-                    "
-hypotheses:
-{}",
-                    format_id_list(&payload.hypothesis_ids, 20)
-                ));
-            }
-            lines.join(
-                "
-",
-            )
-        }
-        CapabilityResponse::IbcAdvance(payload) => {
-            let mut lines = vec![
-                "# ibc_advance".to_string(),
-                format!("advanced: {}", payload.advanced),
-                format!("namespace: {}", payload.namespace),
-                format!(
-                    "pc: {}  status: {}  epoch: {}",
-                    payload.pc, payload.status, payload.epoch
-                ),
-                format!("note: {}", payload.note),
-                format_agent_brief_section(&payload.agent_brief),
-            ];
-            if !payload.hypothesis_ids.is_empty() {
-                lines.push(format!(
-                    "
-hypotheses:
-{}",
-                    format_id_list(&payload.hypothesis_ids, 20)
-                ));
-            }
-            lines.join(
-                "
-",
-            )
-        }
-        CapabilityResponse::HypothesisCreate(payload) => format!(
-            "# hypothesis_create\nid: {}\ntitle: {}\nnotes:\n{}\nevidence:\n{}",
-            payload.hypothesis.id,
-            payload.hypothesis.title,
-            truncate_chars(&payload.hypothesis.notes, 2_000),
-            format_id_list(&payload.hypothesis.evidence_ids, 20)
-        ),
-        CapabilityResponse::HypothesisUpdate(payload) => format!(
-            "# hypothesis_update\nid: {}\ntitle: {}\nnotes:\n{}\nevidence:\n{}",
-            payload.hypothesis.id,
-            payload.hypothesis.title,
-            truncate_chars(&payload.hypothesis.notes, 2_000),
-            format_id_list(&payload.hypothesis.evidence_ids, 20)
-        ),
-        CapabilityResponse::ReportGenerate(payload) => {
-            let mut lines = vec![
-                "# report_generate".to_string(),
-                format!("id: {}", payload.report.id),
-                format!("topic: {}", payload.report.topic),
-                format!(
-                    "\n## Body\n{}",
-                    truncate_chars(&payload.report.body, 10_000)
-                ),
-            ];
-            if !payload.report.evidence_ids.is_empty() {
-                lines.push(format!(
-                    "\n## Evidence IDs\n{}",
-                    format_id_list(&payload.report.evidence_ids, 20)
-                ));
-            }
-            if let Some(artifact) = &payload.artifact {
-                lines.push(format!("\nartifact: {}", artifact.relative_path));
-            }
-            lines.join("\n")
-        }
-        CapabilityResponse::TraceImport(payload) => format!(
-            "# trace_import\nimported: {}\nevidence: {}\nevidence_ids:\n{}",
-            payload.imported,
-            payload.evidence_count,
-            format_id_list(&payload.evidence_ids, 20)
-        ),
-        CapabilityResponse::TraceQuery(payload) => {
-            let mut lines = vec![
-                "# trace_query".to_string(),
-                format!("events: {}", payload.events.len()),
-            ];
-            for event in payload.events.iter().take(40) {
-                lines.push(format!(
-                    "- {}  {}/{}  kind={}  loc={:?}  {}",
-                    event.timestamp,
-                    event.process,
-                    event.thread,
-                    event.kind,
-                    event.location,
-                    truncate_chars(&event.payload.to_string(), 160)
-                ));
-            }
-            if payload.events.len() > 40 {
-                lines.push(format!("- ... {} more", payload.events.len() - 40));
-            }
-            lines.join("\n")
-        }
-    };
-    truncate_chars(&body, MAX_CHARS)
-}
-
-fn truncate_chars(input: &str, max_chars: usize) -> String {
-    if input.chars().count() <= max_chars {
-        return input.to_string();
-    }
-    let mut out = input.chars().take(max_chars).collect::<String>();
-    out.push_str("\n...[truncated]");
-    out
-}
-
-fn format_id_list(ids: &[String], limit: usize) -> String {
-    if ids.is_empty() {
-        return "- none".to_string();
-    }
-    let mut lines = ids
-        .iter()
-        .take(limit)
-        .map(|id| format!("- {id}"))
-        .collect::<Vec<_>>();
-    if ids.len() > limit {
-        lines.push(format!("- ... {} more", ids.len() - limit));
-    }
-    lines.join("\n")
-}
-
-fn format_analysis_summary(summary: &revx_core::AnalysisSummary) -> String {
-    let mut out = format!(
-        "binary_id: {}\nformat: {:?}\narch: {:?}\nfunctions: {}\nimports: {}\nexports: {}\nstrings: {}\ntyped_functions: {}\nstructured_pseudocode: {}\nevidence: {}\ndebug: {:?} types={} fn_hints={} var_hints={}",
-        summary.binary_id,
-        summary.format,
-        summary.architecture,
-        summary.function_count,
-        summary.import_count,
-        summary.export_count,
-        summary.string_count,
-        summary.typed_function_count,
-        summary.structured_pseudocode_count,
-        summary.evidence_count,
-        summary.debug_import_coverage.status,
-        summary.debug_import_coverage.imported_type_count,
-        summary.debug_import_coverage.imported_function_hint_count,
-        summary.debug_import_coverage.imported_variable_hint_count
-    );
-    if !summary.warnings.is_empty() {
-        out.push_str("\nwarnings:");
-        for warning in summary.warnings.iter().take(8) {
-            out.push_str(&format!("\n- {warning}"));
-        }
-    }
-    out
-}
-
-fn format_agent_brief_section(brief: &AgentInteractionBrief) -> String {
-    let mut lines = vec![
-        "\n## Agent Brief".to_string(),
-        format!(
-            "headline: {}",
-            if brief.headline.is_empty() {
-                "-"
-            } else {
-                &brief.headline
-            }
-        ),
-    ];
-    if let Some(lattice) = &brief.semantic_lattice {
-        lines.push(revx_analysis::format_semantic_lattice(lattice));
-    }
-    if !brief.key_findings.is_empty() {
-        lines.push("key_findings:".to_string());
-        for item in brief.key_findings.iter().take(12) {
-            lines.push(format!("- {}", truncate_chars(item, 240)));
-        }
-    }
-    if !brief.open_questions.is_empty() {
-        lines.push("open_questions:".to_string());
-        for item in brief.open_questions.iter().take(8) {
-            lines.push(format!("- {}", truncate_chars(item, 240)));
-        }
-    }
-    if !brief.next_actions.is_empty() {
-        lines.push("next_actions:".to_string());
-        for action in brief.next_actions.iter().take(8) {
-            lines.push(format!(
-                "- p{} `{}`{}: {}\n  args: {}",
-                action.priority,
-                action.tool,
-                action
-                    .label
-                    .as_deref()
-                    .map(|label| format!(" ({label})"))
-                    .unwrap_or_default(),
-                truncate_chars(&action.reason, 200),
-                truncate_chars(&action.args.to_string(), 300)
-            ));
-        }
-        if let Some(top) = brief.next_actions.first() {
-            lines.push(format!(
-                "EXECUTE NOW: `{}` args={}",
-                top.tool,
-                truncate_chars(&top.args.to_string(), 400)
-            ));
-        }
-    }
-    if !brief.stop_conditions.is_empty() {
-        lines.push("stop_conditions:".to_string());
-        for item in brief.stop_conditions.iter().take(6) {
-            lines.push(format!("- {item}"));
-        }
-    }
-    lines.join("\n")
-}
-
-fn render_function_profile(payload: &FunctionProfileResponse) -> String {
-    let function = &payload.function;
-    let mut lines = vec![
-        "# function_profile".to_string(),
-        format!("name: {}", function.name),
-        format!("address: 0x{:x}", function.address),
-        format!("size: {}", function.size),
-        format!(
-            "blocks: {}  callers: {}  callees: {}  xrefs_in: {}  xrefs_out: {}",
-            function.blocks.len(),
-            payload.callers.len(),
-            payload.callees.len(),
-            payload.incoming_xrefs.len(),
-            payload.outgoing_xrefs.len()
-        ),
-    ];
-    if let Some(unit) = &function.pseudocode {
-        if let Some(lattice) = &unit.semantic_lattice {
-            lines.push(String::new());
-            lines.push(revx_analysis::format_semantic_lattice(lattice));
-        }
-    } else if let Some(lattice) = &payload.agent_brief.semantic_lattice {
-        lines.push(String::new());
-        lines.push(revx_analysis::format_semantic_lattice(lattice));
-    }
-    if let Some(stack) = &function.stack_summary {
-        lines.push(format!(
-            "stack: frame={:?} cc={:?} ret={:?} stack_args={:?}",
-            stack.frame_size, stack.calling_convention, stack.return_type, stack.stack_arg_bytes
-        ));
-    }
-    if !function.arguments.is_empty() {
-        lines.push("\n## Arguments".to_string());
-        for arg in function.arguments.iter().take(16) {
-            lines.push(format!(
-                "- {}  {:?}  {:?}  type={}  conf={:.2}  @{}",
-                arg.name,
-                arg.role,
-                arg.storage,
-                arg.type_name.as_deref().unwrap_or("?"),
-                arg.confidence,
-                arg.location
-            ));
-        }
-    }
-    if !function.locals.is_empty() {
-        lines.push("\n## Locals".to_string());
-        for local in function.locals.iter().take(24) {
-            lines.push(format!(
-                "- {}  {:?}  type={}  conf={:.2}  @{}",
-                local.name,
-                local.storage,
-                local.type_name.as_deref().unwrap_or("?"),
-                local.confidence,
-                local.location
-            ));
-        }
-    }
-    if !payload.callers.is_empty() {
-        lines.push("\n## Callers".to_string());
-        for edge in payload.callers.iter().take(24) {
-            lines.push(format!(
-                "- {} (0x{:x}) -[{}]-> {} (0x{:x})",
-                edge.caller_name,
-                edge.caller_address,
-                edge.kind,
-                edge.callee_name.as_deref().unwrap_or(&function.name),
-                edge.callee_address
-            ));
-        }
-        if payload.callers.len() > 24 {
-            lines.push(format!("- ... {} more", payload.callers.len() - 24));
-        }
-    }
-    if !payload.callees.is_empty() {
-        lines.push("\n## Callees".to_string());
-        for edge in payload.callees.iter().take(24) {
-            lines.push(format!(
-                "- {} (0x{:x}) -[{}]-> {} (0x{:x})",
-                edge.caller_name,
-                edge.caller_address,
-                edge.kind,
-                edge.callee_name.as_deref().unwrap_or("?"),
-                edge.callee_address
-            ));
-        }
-        if payload.callees.len() > 24 {
-            lines.push(format!("- ... {} more", payload.callees.len() - 24));
-        }
-    }
-    if !payload.incoming_xrefs.is_empty() {
-        lines.push("\n## Incoming xrefs".to_string());
-        for reference in payload.incoming_xrefs.iter().take(30) {
-            lines.push(format!(
-                "- 0x{:x} -> 0x{:x}  {}",
-                reference.from, reference.to, reference.kind
-            ));
-        }
-        if payload.incoming_xrefs.len() > 30 {
-            lines.push(format!("- ... {} more", payload.incoming_xrefs.len() - 30));
-        }
-    }
-    if !payload.outgoing_xrefs.is_empty() {
-        lines.push("\n## Outgoing xrefs".to_string());
-        for reference in payload.outgoing_xrefs.iter().take(30) {
-            lines.push(format!(
-                "- 0x{:x} -> 0x{:x}  {}",
-                reference.from, reference.to, reference.kind
-            ));
-        }
-        if payload.outgoing_xrefs.len() > 30 {
-            lines.push(format!("- ... {} more", payload.outgoing_xrefs.len() - 30));
-        }
-    }
-    lines.push(format!(
-        "\n## Digest\n{}",
-        function_pseudocode_digest(
-            function.pseudocode.as_ref(),
-            &payload.callees,
-            &payload.callers
-        )
-    ));
-    if let Some(unit) = &function.pseudocode {
-        lines.push("\n## Pseudocode".to_string());
-        lines.push(format!(
-            "```{}\n{}\n```",
-            unit.language,
-            truncate_chars(&unit.text, 6_000)
-        ));
-    }
-    if !function.warnings.is_empty() {
-        lines.push("\n## Warnings".to_string());
-        for warning in function.warnings.iter().take(12) {
-            lines.push(format!("- {warning}"));
-        }
-    }
-    if !function.evidence_ids.is_empty() {
-        lines.push(format!(
-            "\n## Evidence IDs\n{}",
-            format_id_list(&function.evidence_ids, 16)
-        ));
-    }
-    if !payload.agent_brief.headline.is_empty() || !payload.agent_brief.next_actions.is_empty() {
-        lines.push(format_agent_brief_section(&payload.agent_brief));
-    } else {
-        lines.push(format!(
-            "\nnext: decompile_function({}) | disassemble_function({}) | xrefs_query(0x{:x})",
-            function.name, function.name, function.address
-        ));
-    }
-    if let Some(artifact) = &payload.artifact {
-        lines.push(format!("artifact: {}", artifact.relative_path));
-    }
-    lines.join("\n")
-}
-
-fn render_disassembly(payload: &DisassembleFunctionResponse) -> String {
-    let mut lines = vec![
-        "# disassemble_function".to_string(),
-        format!("name: {}", payload.function_name),
-        format!("address: 0x{:x}", payload.address),
-        format!("blocks: {}", payload.blocks.len()),
-    ];
-    let mut shown_insns = 0usize;
-    const MAX_INSNS: usize = 220;
-    for (index, block) in payload.blocks.iter().enumerate() {
-        if shown_insns >= MAX_INSNS {
-            lines.push(format!(
-                "\n... truncated remaining blocks ({} total)",
-                payload.blocks.len()
-            ));
-            break;
-        }
-        lines.push(format!(
-            "\n## Block {}  0x{:x}  size={}",
-            index, block.address, block.size
-        ));
-        for insn in &block.instructions {
-            if shown_insns >= MAX_INSNS {
-                lines.push(format!(
-                    "... truncated ({} more insns in this block)",
-                    block.instructions.len().saturating_sub(
-                        block
-                            .instructions
-                            .iter()
-                            .position(|item| item.address == insn.address)
-                            .unwrap_or(0)
-                    )
-                ));
-                break;
-            }
-            lines.push(format!(
-                "0x{:x}:  {:<16}  {}",
-                insn.address, insn.bytes, insn.text
-            ));
-            shown_insns += 1;
-        }
-    }
-    if let Some(annotations) = &payload.annotations {
-        lines.push(format!(
-            "\nannotations_artifact: {}",
-            annotations.relative_path
-        ));
-    }
-    if let Some(artifact) = &payload.artifact {
-        lines.push(format!("artifact: {}", artifact.relative_path));
-    }
-    lines.push(format!(
-        "\nnext: decompile_function({}) | function_profile({})",
-        payload.function_name, payload.function_name
-    ));
-    lines.join("\n")
-}
-
-fn mcp_tools_manifest() -> Vec<serde_json::Value> {
-    vec![
-        tool_manifest(
-            "project_open",
-            "Open a revx workspace",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "path": { "type": "string" } },
-                "required": ["path"]
-            }),
-        ),
-        tool_manifest(
-            "project_status",
-            "Read workspace status",
-            serde_json::json!({
-                "type": "object",
-                "properties": {}
-            }),
-        ),
-        tool_manifest(
-            "object_identify",
-            "Identify an arbitrary file or directory as a universal object graph",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "path": { "type": "string" },
-                    "max_depth": { "type": ["integer", "null"] },
-                    "max_children": { "type": ["integer", "null"] },
-                    "include_graph": { "type": ["boolean", "null"] }
-                },
-                "required": ["path"]
-            }),
-        ),
-        tool_manifest(
-            "object_search",
-            "Search persisted universal objects by id, path, format, hash, metadata, or analyzer output",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": "string" },
-                    "kind": {
-                        "type": ["string", "null"],
-                        "enum": [
-                            "file", "directory", "archive", "binary", "text", "image",
-                            "document", "package", "filesystem_image", "memory_dump",
-                            "network_capture", "database", "model", "unknown", null
-                        ]
-                    },
-                    "limit": { "type": ["integer", "null"] }
-                },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "object_profile",
-            "Read a persisted universal object profile with graph edges and evidence ids",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "query": { "type": "string" } },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "object_materialize",
-            "Materialize a persisted object, including virtual container children, into an artifact",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": "string" },
-                    "preview_bytes": { "type": ["integer", "null"] }
-                },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "object_extract_range",
-            "Extract a byte range from a persisted object, including virtual container children, into a new evidence artifact",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": "string" },
-                    "offset": { "type": "integer" },
-                    "length": { "type": "integer" },
-                    "context_bytes": { "type": ["integer", "null"] },
-                    "preview_bytes": { "type": ["integer", "null"] }
-                },
-                "required": ["query", "offset", "length"]
-            }),
-        ),
-        tool_manifest(
-            "object_scan_signatures",
-            "Scan a persisted object for embedded file signatures and offsets that can be extracted as follow-up evidence",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": "string" },
-                    "limit": { "type": ["integer", "null"] },
-                    "max_object_bytes": { "type": ["integer", "null"] },
-                    "preview_bytes": { "type": ["integer", "null"] }
-                },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "object_carve_signatures",
-            "Scan a persisted object for bounded embedded signatures and carve them into evidence artifacts",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": "string" },
-                    "limit": { "type": ["integer", "null"] },
-                    "max_object_bytes": { "type": ["integer", "null"] },
-                    "max_carve_bytes": { "type": ["integer", "null"] },
-                    "min_confidence": { "type": ["number", "null"] },
-                    "preview_bytes": { "type": ["integer", "null"] }
-                },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "object_carve_identify",
-            "Carve bounded embedded signatures into artifacts, recursively identify each carved artifact, and persist the resulting object graphs",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": "string" },
-                    "limit": { "type": ["integer", "null"] },
-                    "max_object_bytes": { "type": ["integer", "null"] },
-                    "max_carve_bytes": { "type": ["integer", "null"] },
-                    "min_confidence": { "type": ["number", "null"] },
-                    "preview_bytes": { "type": ["integer", "null"] },
-                    "max_depth": { "type": ["integer", "null"] },
-                    "max_children": { "type": ["integer", "null"] }
-                },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "object_analyze",
-            "Analyze a persisted object. Read agent_brief.headline + next_actions[0] and execute exactly one ranked follow-up using next_actions[0].args; honor stop_conditions",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": "string" },
-                    "analyzers": {
-                        "type": ["array", "null"],
-                        "items": {
-                            "type": "string",
-                            "enum": ["auto", "byte_histogram", "strings", "structured_text", "zip_listing", "android_package", "dex_bytecode", "ios_package", "java_archive", "jvm_class", "python_bytecode", "shell_link", "portable_executable", "dotnet_metadata", "elf_binary", "macho_binary", "open_xml_document", "sqlite_schema", "wasm_module", "pdf_document", "png_image", "jpeg_image", "gif_image", "bmp_image", "riff_container", "pcap_capture", "ole_compound", "safe_tensors_model", "gguf_model", "pytorch_model", "iso_bmff", "cab_archive", "ar_archive", "font_file", "tiff_image", "audio_media", "disk_image", "unknown_blob"]
-                        }
-                    }
-                },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "object_plugin_list",
-            "List workspace object analyzer plugins from .revx/plugins/*.json",
-            serde_json::json!({
-                "type": "object",
-                "properties": {}
-            }),
-        ),
-        tool_manifest(
-            "object_plugin_run",
-            "Materialize an object, run a workspace plugin analyzer command, and persist its output as evidence",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "plugin_id": { "type": "string" },
-                    "query": { "type": "string" },
-                    "timeout_ms": { "type": ["integer", "null"] }
-                },
-                "required": ["plugin_id", "query"]
-            }),
-        ),
-        tool_manifest(
-            "object_register_binary",
-            "Materialize an object and register it as a binary survey",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "query": { "type": "string" } },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "object_analyze_binary",
-            "Materialize an object and run binary analysis over it",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": "string" },
-                    "profile": { "type": "string", "enum": ["fast", "full"] }
-                },
-                "required": ["query", "profile"]
-            }),
-        ),
-        tool_manifest(
-            "object_pipeline",
-            "Run recursive object discovery/carving. Consume agent_brief.next_actions[0] for the single next high-value tool call",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "path": { "type": "string" },
-                    "max_depth": { "type": ["integer", "null"] },
-                    "max_children": { "type": ["integer", "null"] },
-                    "object_limit": { "type": ["integer", "null"] },
-                    "analyze_objects": { "type": ["boolean", "null"] },
-                    "carve_embedded": { "type": ["boolean", "null"] },
-                    "carve_limit": { "type": ["integer", "null"] },
-                    "max_carve_object_bytes": { "type": ["integer", "null"] },
-                    "max_carve_bytes": { "type": ["integer", "null"] },
-                    "min_carve_confidence": { "type": ["number", "null"] },
-                    "carve_max_depth": { "type": ["integer", "null"] },
-                    "carve_max_children": { "type": ["integer", "null"] },
-                    "plugin_ids": { "type": ["array", "null"], "items": { "type": "string" } },
-                    "analyze_binaries": { "type": ["boolean", "null"] },
-                    "binary_profile": { "type": ["string", "null"], "enum": ["fast", "full", null] }
-                },
-                "required": ["path"]
-            }),
-        ),
-        tool_manifest(
-            "binary_list",
-            "List registered binaries",
-            serde_json::json!({
-                "type": "object",
-                "properties": {}
-            }),
-        ),
-        tool_manifest(
-            "analysis_run",
-            "Analyze a binary and return run status plus function/import/string coverage in tool text",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "binary_path": { "type": "string" },
-                    "profile": { "type": "string", "enum": ["fast", "full"] }
-                },
-                "required": ["binary_path", "profile"]
-            }),
-        ),
-        tool_manifest(
-            "analysis_status",
-            "Read analysis run status",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "run_id": { "type": ["string", "null"] } }
-            }),
-        ),
-        tool_manifest(
-            "binary_survey",
-            "Read binary survey stats (format/arch/functions/strings/debug coverage) with next-step guidance",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "binary_id": { "type": ["string", "null"] } }
-            }),
-        ),
-        tool_manifest(
-            "function_search",
-            "Search functions by name/address. Text includes ranked matches with addresses and sizes; use function_profile/decompile_function on a hit",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "query": { "type": "string" } },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "function_profile",
-            "Read a function dossier: args/locals, callers/callees, xrefs, and pseudocode preview in the tool text response",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "query": { "type": "string" } },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "decompile_cache_status",
-            "List function pseudocode artifact and per-strategy cache entries (fast/full/hotblock) for a function query",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "query": { "type": "string" } },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "decompile_function",
-            "Return deterministic pseudocode text and region outline for a function. strategy: auto|cached|fast|full|hotblock; force_refresh recomputes",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": "string" },
-                    "strategy": {
-                        "type": "string",
-                        "enum": ["auto", "cached", "fast", "full", "hotblock"]
-                    },
-                    "force_refresh": { "type": "boolean" }
-                },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "disassemble_function",
-            "Return recovered basic blocks and instruction listing in tool text (bounded for large functions)",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "query": { "type": "string" } },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "xrefs_query",
-            "Query xrefs; tool text lists from->to with kind. Prefer address queries from string/function hits",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "query": { "type": "string" } },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "callgraph_slice",
-            "Return callgraph edges around a function; tool text lists caller/callee names and addresses",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "query": { "type": "string" } },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "string_search",
-            "Search recovered strings; text lists address+value matches. Follow with xrefs_query on a string address",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "pattern": { "type": "string" } },
-                "required": ["pattern"]
-            }),
-        ),
-        tool_manifest(
-            "search_bytes",
-            "Search bytes in the latest analyzed binary",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "pattern": { "type": "string" } },
-                "required": ["pattern"]
-            }),
-        ),
-        tool_manifest(
-            "object_search_content",
-            "Search text or hex bytes across persisted universal objects, including virtual container children",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "pattern": { "type": "string" },
-                    "mode": { "type": ["string", "null"], "enum": ["text", "hex", null] },
-                    "query": { "type": ["string", "null"] },
-                    "limit": { "type": ["integer", "null"] },
-                    "per_object_limit": { "type": ["integer", "null"] },
-                    "max_object_bytes": { "type": ["integer", "null"] }
-                },
-                "required": ["pattern"]
-            }),
-        ),
-        tool_manifest(
-            "artifact_read",
-            "Read a bounded preview of a workspace artifact by relative path or blake3 hash",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "relative_path": { "type": ["string", "null"] },
-                    "hash_blake3": { "type": ["string", "null"] },
-                    "offset": { "type": ["integer", "null"] },
-                    "max_bytes": { "type": ["integer", "null"] }
-                }
-            }),
-        ),
-        tool_manifest(
-            "artifact_list",
-            "List and search workspace artifacts with roles and provenance references for agent navigation",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": ["string", "null"] },
-                    "content_type": { "type": ["string", "null"] },
-                    "role": { "type": ["string", "null"] },
-                    "limit": { "type": ["integer", "null"] },
-                    "include_unreferenced": { "type": ["boolean", "null"] }
-                }
-            }),
-        ),
-        tool_manifest(
-            "evidence_pack",
-            "Read evidence for a subject; tool text includes summaries and key details, not just counts",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "subject": { "type": "string" } },
-                "required": ["subject"]
-            }),
-        ),
-        tool_manifest(
-            "evidence_graph",
-            "Derive a bounded evidence graph connecting subjects, objects, artifacts, provenance, binaries, and functions",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "subject": { "type": "string" },
-                    "depth": { "type": ["integer", "null"] },
-                    "limit": { "type": ["integer", "null"] }
-                },
-                "required": ["subject"]
-            }),
-        ),
-        tool_manifest(
-            "symbolic_solve",
-            "Solve a finite symbolic constraint case and persist the result as evidence",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "subject": { "type": "string" },
-                    "variables": {
-                        "type": "array",
-                        "items": {
-                            "type": "object",
-                            "properties": {
-                                "name": { "type": "string" },
-                                "domain": {
-                                    "type": "object",
-                                    "properties": {
-                                        "kind": { "type": "string", "enum": ["int_range", "int_values"] },
-                                        "min": { "type": "integer" },
-                                        "max": { "type": "integer" },
-                                        "values": { "type": "array", "items": { "type": "integer" } }
-                                    },
-                                    "required": ["kind"]
-                                }
-                            },
-                            "required": ["name", "domain"]
-                        }
-                    },
-                    "constraints": {
-                        "type": "array",
-                        "items": {
-                            "type": "object",
-                            "properties": {
-                                "id": { "type": ["string", "null"] },
-                                "left": {
-                                    "type": "object",
-                                    "properties": {
-                                        "terms": {
-                                            "type": "array",
-                                            "items": {
-                                                "type": "object",
-                                                "properties": {
-                                                    "variable": { "type": "string" },
-                                                    "coefficient": { "type": "integer" }
-                                                },
-                                                "required": ["variable", "coefficient"]
-                                            }
-                                        },
-                                        "constant": { "type": "integer" }
-                                    }
-                                },
-                                "op": { "type": "string", "enum": ["eq", "ne", "lt", "le", "gt", "ge"] },
-                                "right": {
-                                    "type": "object",
-                                    "properties": {
-                                        "terms": {
-                                            "type": "array",
-                                            "items": {
-                                                "type": "object",
-                                                "properties": {
-                                                    "variable": { "type": "string" },
-                                                    "coefficient": { "type": "integer" }
-                                                },
-                                                "required": ["variable", "coefficient"]
-                                            }
-                                        },
-                                        "constant": { "type": "integer" }
-                                    }
-                                }
-                            },
-                            "required": ["left", "op", "right"]
-                        }
-                    },
-                    "max_solutions": { "type": ["integer", "null"] },
-                    "iteration_limit": { "type": ["integer", "null"] }
-                },
-                "required": ["subject", "variables", "constraints"]
-            }),
-        ),
-        tool_manifest(
-            "analysis_brief",
-            "One-shot multi-hop RE brief for agents: ranked strings/functions, xref-backed hot functions, pseudocode previews, and next_actions with args. Prefer this over chaining string_search→xrefs→function_profile manually",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": "string" },
-                    "string_limit": { "type": ["integer", "null"] },
-                    "function_limit": { "type": ["integer", "null"] },
-                    "hot_function_limit": { "type": ["integer", "null"] },
-                    "xref_limit": { "type": ["integer", "null"] },
-                    "include_pseudocode": { "type": ["boolean", "null"] }
-                },
-                "required": ["query"]
-            }),
-        ),
-        tool_manifest(
-            "investigation_run",
-            "Run an AI-native investigation. Tool text includes agent_brief, ranked next_actions with args, key findings, report body; execute next_actions[0] only then reassess",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "subject": { "type": "string" },
-                    "path": { "type": ["string", "null"] },
-                    "run_object_pipeline": { "type": ["boolean", "null"] },
-                    "max_depth": { "type": ["integer", "null"] },
-                    "max_children": { "type": ["integer", "null"] },
-                    "object_limit": { "type": ["integer", "null"] },
-                    "carve_max_depth": { "type": ["integer", "null"] },
-                    "carve_max_children": { "type": ["integer", "null"] },
-                    "plugin_ids": { "type": ["array", "null"], "items": { "type": "string" } },
-                    "analyze_binaries": { "type": ["boolean", "null"] },
-                    "binary_profile": { "type": ["string", "null"], "enum": ["fast", "full", null] },
-                    "graph_depth": { "type": ["integer", "null"] },
-                    "graph_limit": { "type": ["integer", "null"] },
-                    "trace_kind": { "type": ["string", "null"] },
-                    "trace_limit": { "type": ["integer", "null"] }
-                },
-                "required": ["subject"]
-            }),
-        ),
-        tool_manifest(
-            "ibc_status",
-            "Inspect durable CASL IBC continuum: pc/status/epoch/witnesses/orbit hypotheses and ranked next_actions",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": ["string", "null"] }
-                }
-            }),
-        ),
-        tool_manifest(
-            "ibc_advance",
-            "Advance durable CASL IBC continuum: force_next or warp by tool+query; auto-binds orbit hypotheses",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": ["string", "null"] },
-                    "tool": { "type": ["string", "null"] },
-                    "query": { "type": ["string", "null"] },
-                    "force_next": { "type": ["boolean", "null"] }
-                }
-            }),
-        ),
-        tool_manifest(
-            "hypothesis_create",
-            "Create a workspace-local hypothesis",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "title": { "type": "string" },
-                    "notes": { "type": "string" },
-                    "evidence_ids": { "type": "array", "items": { "type": "string" } }
-                },
-                "required": ["title", "notes", "evidence_ids"]
-            }),
-        ),
-        tool_manifest(
-            "hypothesis_update",
-            "Update a workspace-local hypothesis",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "id": { "type": "string" },
-                    "title": { "type": ["string", "null"] },
-                    "notes": { "type": ["string", "null"] },
-                    "evidence_ids": { "type": ["array", "null"], "items": { "type": "string" } }
-                },
-                "required": ["id"]
-            }),
-        ),
-        tool_manifest(
-            "report_generate",
-            "Generate a report preview and artifact",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "topic": { "type": "string" } },
-                "required": ["topic"]
-            }),
-        ),
-        tool_manifest(
-            "trace_import",
-            "Import trace events",
-            serde_json::json!({
-                "type": "object",
-                "properties": { "events": { "type": "array" } },
-                "required": ["events"]
-            }),
-        ),
-        tool_manifest(
-            "trace_query",
-            "Query imported traces",
-            serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "kind": { "type": ["string", "null"] },
-                    "limit": { "type": ["integer", "null"] }
-                }
-            }),
-        ),
-    ]
-}
-
-fn tool_manifest(
-    name: &str,
-    description: &str,
-    input_schema: serde_json::Value,
-) -> serde_json::Value {
-    serde_json::json!({
-        "name": name,
-        "description": description,
-        "inputSchema": input_schema,
     })
 }
